TSNERunner is created using factory so it would be easier to mock it
