+++ src/interface/VizBin/src/lu/uni/lcsb/vizbin/DataSetUtils.java
@@ -11,6 +11,7 @@ import lu.uni.lcsb.vizbin.data.DataSet;
 import lu.uni.lcsb.vizbin.data.Sequence;
 import lu.uni.lcsb.vizbin.pca.IPrincipleComponentAnalysis;
 import lu.uni.lcsb.vizbin.pca.PcaType;
+import lu.uni.lcsb.vizbin.tsne.*;
 import no.uib.cipr.matrix.NotConvergedException;
 
 public class DataSetUtils {
@@ -40,6 +41,8 @@ public class DataSetUtils {
 	// private static ArrayList<Sequence> pointList = null;
 	// private static List<Point2D> polygonPoints = null;
 	private static JFrame		drawingFrame			= null;
+
+    static TSNERunnerFactory tsneRunnerFactory = new TSNERunnerFactoryImpl();
 
 	public static void createKmers(DataSet dataSet, int k, boolean mergeRevComp, ProcessGuiParameters guiParams) {
 		int totalKmers = 0;
@@ -262,73 +265,10 @@ public class DataSetUtils {
 
 		logger.debug("Running command: \"" + tsneCmd + "\" in directory: " + dir + "\n" + "Number of threads: " + numThreads + "\n" + "Seed: " + seed);
 
-		TSNERunner tsne = new TSNERunner(tsneCmd, dir, guiParameters);
+		TSNERunner tsne = tsneRunnerFactory.createRunner(tsneCmd, dir, guiParameters);
 		Thread tr = new Thread(tsne, "TSNERunner");
 		tr.start();
 		tr.join();
-	}
-
-	static class TSNERunner implements Runnable {
-		private File									command;
-		private String								dir;
-		private ProcessGuiParameters	guiParameters;
-
-		public TSNERunner(File command, String dir, ProcessGuiParameters guiParameters) {
-			this.command = command;
-			this.dir = dir;
-			this.guiParameters = guiParameters;
-		}
-
-		public void run() {
-			Process process;
-			try {
-				process = Runtime.getRuntime().exec(new String[] { command.getAbsolutePath() }, new String[] {}, new File(dir));
-				/*
-				 * Tomasz Sternal - removed this line, it was not updating progress bar
-				 * status in mainFrame in real time
-				 *
-				 * SwingUtilities.invokeAndWait(new
-				 * ProgressReader(process.getInputStream(), label_status, progBar));
-				 */
-				BufferedReader reader = new BufferedReader(new InputStreamReader(process.getInputStream()));
-				String line = reader.readLine();
-				Integer newProgress = 0;
-				while (line != null) {
-					logger.debug("TSNE: " + line);
-
-					// other operations take us to 35% completion, split
-					// remaining 65% between TSNE phases
-					if (line.startsWith("Building tree"))
-						newProgress = 5; // 40%
-					if (line.startsWith("Learning embedding"))
-						newProgress = 5; // 45%
-					if (line.startsWith("Iteration")) { // there are 20
-						// iterations
-						newProgress = 2; // add 2% progress per iteration
-						String logEntry = "T-SNE: " + line.substring(0, line.indexOf(':')) + "/1000";
-						if (guiParameters != null) {
-							guiParameters.getStatusLabel().setText(logEntry);
-						} else {
-							logger.debug("[PROGRESS BAR] " + logEntry);
-						}
-					}
-					if (line.contains("Wrote the")) {
-						newProgress = 5; // 90%
-					}
-					if (guiParameters != null) {
-						guiParameters.getProgessBar().setValue(guiParameters.getProgessBar().getValue() + newProgress);
-					} else {
-						// logger.debug("[PROGRESS BAR] " + "next " + newProgress + "% ");
-
-					}
-					line = reader.readLine();
-				}
-				reader.close();
-				process.waitFor();
-			} catch (Exception e) {
-				e.printStackTrace();
-			}
-		}
 	}
 
 	public static boolean isIsDataSetCreated() {
