+++ src/interface/VizBin/src/lu/uni/lcsb/vizbin/tsne/TSNERunner.java
@@ -1,0 +1,7 @@
+package lu.uni.lcsb.vizbin.tsne;
+
+public interface TSNERunner extends Runnable {
+
+  void run();
+
+}
